update profile settings
